refactor: moved useAppSettings
